+++ app.py
@@ -1,91 +1,24 @@
-from fastapi import FastAPI, HTTPException, Query
+from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from pydantic import BaseModel
-import requests
-import json
-# from db_control import crud, mymodels
-from db_control import crud, mymodels
-# MySQLのテーブル作成
-# from db_control.create_tables import init_db
 
-# # アプリケーション初期化時にテーブルを作成
-# init_db()
-
-
-class Customer(BaseModel):
-    customer_id: str
-    customer_name: str
-    age: int
-    gender: str
-
+# 新しいルーター（/answers）をインポート
+from routers import answer
 
 app = FastAPI()
 
-# CORSミドルウェアの設定
+# CORSミドルウェアの設定（ローカル開発用）
 app.add_middleware(
     CORSMiddleware,
-    allow_origins=["*"],
+    allow_origins=["*"],  # 必要に応じて制限可能
     allow_credentials=True,
     allow_methods=["*"],
     allow_headers=["*"],
 )
 
+# トップページ（動作確認用）
 @app.get("/")
 def index():
-    return {"message": "FastAPI top page!"}
+    return {"message": "FastAPI Zukiraku API running!"}
 
-
-@app.post("/customers")
-def create_customer(customer: Customer):
-    values = customer.dict()
-    tmp = crud.myinsert(mymodels.Customers, values)
-    result = crud.myselect(mymodels.Customers, values.get("customer_id"))
-
-    if result:
-        result_obj = json.loads(result)
-        return result_obj if result_obj else None
-    return None
-
-
-@app.get("/customers")
-def read_one_customer(customer_id: str = Query(...)):
-    result = crud.myselect(mymodels.Customers, customer_id)
-    if not result:
-        raise HTTPException(status_code=404, detail="Customer not found")
-    result_obj = json.loads(result)
-    return result_obj[0] if result_obj else None
-
-@app.get("/allcustomers")
-def read_all_customer():
-    result = crud.myselectAll(mymodels.Customers)
-    # 結果がNoneの場合は空配列を返す
-    if not result:
-        return []
-    # JSON文字列をPythonオブジェクトに変換
-    return json.loads(result)
-
-
-@app.put("/customers")
-def update_customer(customer: Customer):
-    values = customer.dict()
-    values_original = values.copy()
-    tmp = crud.myupdate(mymodels.Customers, values)
-    result = crud.myselect(mymodels.Customers, values_original.get("customer_id"))
-    if not result:
-        raise HTTPException(status_code=404, detail="Customer not found")
-    result_obj = json.loads(result)
-    return result_obj[0] if result_obj else None
-
-
-@app.delete("/customers")
-def delete_customer(customer_id: str = Query(...)):
-    result = crud.mydelete(mymodels.Customers, customer_id)
-    if not result:
-        raise HTTPException(status_code=404, detail="Customer not found")
-    return {"customer_id": customer_id, "status": "deleted"}
-
-
-@app.get("/fetchtest")
-def fetchtest():
-    response = requests.get('https://jsonplaceholder.typicode.com/users')
-    return response.json()
+# /answersルーターを登録（タイプ判定用API）
+app.include_router(answer.router, prefix="", tags=["Headache Screening"])
